move exception handling in the service layer mostly
