+++ domain/src/main/java/com/healthcare/management/system/domain/UserService.java
@@ -2,6 +2,7 @@ package com.healthcare.management.system.domain;
 
 
 import com.healthcare.management.system.domain.model.DomainUser;
+import org.springframework.http.ResponseEntity;
 
 import java.util.List;
 
@@ -10,5 +11,5 @@ public interface UserService {
     public List<DomainUser> findAll();
     public DomainUser createOrUpdate(DomainUser domainUser);
     public DomainUser findById(int theId);
-    public void deleteById(int theId);
+    public ResponseEntity<Object> deleteById(int theId);
 }
