+++ rest-module/src/main/java/com/healthcare/management/system/rest/controller/UserController.java
@@ -1,5 +1,6 @@
 package com.healthcare.management.system.rest.controller;
 
+import com.healthcare.management.system.domain.exceptions.UserNotFoundException;
 import com.healthcare.management.system.domain.model.DomainUser;
 import com.healthcare.management.system.domain.UserService;
 import com.healthcare.management.system.rest.dto.AddUserDTO;
@@ -7,7 +8,6 @@ import com.healthcare.management.system.rest.dto.UpdateUserDTO;
 import com.healthcare.management.system.rest.mapper.AddUserRestMapper;
 import com.healthcare.management.system.rest.mapper.UpdateUserRestMapper;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -28,20 +28,11 @@ public class UserController {
 
     @GetMapping("/users/{userId}")
     public DomainUser getUser(@PathVariable int userId) {
-        DomainUser domainUser = userService.findById(userId);
-
-        if (domainUser == null) {
-            throw new UserNotFoundException(userId);
-        }
-        return domainUser;
+        return userService.findById(userId);
     }
 
     @PostMapping("/users")
     public AddUserDTO createUser(@RequestBody AddUserDTO userDTO) {
-        // just in case they pass an id in JSON ... set id to 0
-        // this is to force a save of new item ... instead of update
-        // domainUser.setId(0);
-
         return addUserMapper.toDTO(userService.createOrUpdate(addUserMapper.fromDTO(userDTO)));
     }
 
@@ -58,13 +49,6 @@ public class UserController {
 
     @DeleteMapping("/users/{userId}")
     public ResponseEntity deleteUser(@PathVariable int userId) {
-        DomainUser domainUser = userService.findById(userId);
-
-        if (domainUser == null) {
-            throw new UserNotFoundException(userId);
-        }
-
-        userService.deleteById(userId);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        return userService.deleteById(userId);
     }
 }
